requirements.txt
# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
jose==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Database
mongoengine
motor==3.3.2
redis==5.0.1

# Utils
requests==2.31.0
rich==13.7.0
httpx==0.25.0
asyncio==3.4.3
aiohttp==3.9.1

# AI Services
openai==1.3.5

# Logging
loguru==0.7.2